CSS Horizontal Navbar - Example D
Responsive Sidenav Example

This example use media queries to transform the sidenav to a top navigation bar when the screen size is 900px or less.

We have also added a media query for screens that are 400px or less, which will vertically stack and center the navigation links.

You will learn more about media queries and responsive web design later in our CSS Tutorial.