Grid Item - The Order of Items
The Grid Layout allows us to position the items anywhere we like.

The first item in the HTML code does not have to appear as the first item in the grid.

The grid items do not have to be displayed in the same order as they are written in the HTML code.