CSS Animations - Specify the fill-mode For an Animation
CSS animations do not affect an element before the first keyframe is played or after the last keyframe is played. The animation-fill-mode property can override this behavior.

The animation-fill-mode property specifies a style for the target element when the animation is not playing (before it starts, after it ends, or both).

The animation-fill-mode property can have the following values:

none - Default value. Animation will not apply any styles to the element before or after it is executing

forwards - The element will retain the style values that is set by the last keyframe (depends on animation-direction and animation-iteration-count)

backwards - The element will get the style values that is set by the first keyframe (depends on animation-direction), and retain this during the animation-delay period

both - The animation will follow the rules for both forwards and backwards, extending the animation properties in both directions

This example lets the <div> element retain the style values from the last keyframe when the animation ends