The CSS object-fit Property The CSS object-fit property is used to specify how an <img> or <video> should be resized to fit its container.
This property tells the content to fill the container in a variety of ways; such as "preserve that aspect ratio" or "stretch up and take up as much space as possible".

Look at the following image from Paris. This image is 400 pixels wide and 300 pixels high

However, if we style the image above to be half its width (200 pixels) and same height (300 pixels), it will look like this

img {
  width: 200px;
  height: 300px;
}