CSS Box Sizing - With Box Sizing
oth divs are the same size now! Hooray!

With the CSS box-sizing Property

The box-sizing property allows us to include the padding and border in an element's total width and height.

If you set box-sizing: border-box; on an element, padding and border are included in the width and height

box-sizing: border-box;