CSS Box Sizing - Box Sizing All Elements
Since the result of using the box-sizing: border-box; is so much better, many developers want all elements on their pages to work this way.

The code below ensures that all elements are sized in this more intuitive way. Many browsers already use box-sizing: border-box; for many form elements (but not all - which is why inputs and text areas look different at width: 100%;).

Applying this to all elements is safe and wise:

Try to remove the box-sizing property from the style element and look what happens.

Notice that the width of input, textarea, and submit button will go outside of the screen.