CSS Media Queries Examples - Orientation: Portrait / Landscape
Media queries can also be used to change layout of a page depending on the orientation of the browser.

You can have a set of CSS properties that will only apply when the browser window is wider than its height, a so called "Landscape" orientation

Resize the browser window. When the width of this document is larger than the height, the background color is "lightblue", otherwise it is "lightgreen".