CSS Flexbox - Flex Items - The order Property
Use the order property to sort the flex items as you like:

The order property specifies the order of the flex items.

The first flex item in the code does not have to appear as the first item in the layout.

The order value must be a number, default value is 0.