CSS Flexbox - Flex Responsive - Responsive Flexbox
In this example, we change the percentage of flex to create different layouts for different screen sizes.

Resize the browser window to see that the direction changes when the screen size is 800px wide or smaller.

Another way is to change the percentage of the flex property of the flex items to create different layouts for different screen sizes. Note that we also have to include flex-wrap: wrap; on the flex container for this example to work: